Purchased Items

+++ src/main/scala/com/platalytics/HDCustomers.scala
@@ -15,12 +15,12 @@ class HDCustomers {
      
      var purchases = sc.textFile("Purchases.csv").map(line => {
       val parts = line.split(",")
-      (parts(1).toInt, parts(4).toInt)
+      (parts(1).toInt, parts(4).toDouble)
      })
      
      var returns= sc.textFile("Returns.csv").map(line => {
       val parts = line.split(",")
-      (parts(1).toInt, parts(4).toInt)
+      (parts(1).toInt, parts(4).toDouble)
      })
